[frontend] extract mod circuit (#196)
This is not a functional change, just extracting a module for Circuit and the
gang.
