+++ crates/frontend/src/circuits/bignum/mod.rs
@@ -8,7 +8,7 @@
 mod tests;
 
 use crate::{
-	compiler::{CircuitBuilder, Wire, WitnessFiller},
+	compiler::{CircuitBuilder, Wire, circuit::WitnessFiller},
 	word::Word,
 };
