+++ crates/frontend/src/circuits/jwt_claims.rs
@@ -1,6 +1,6 @@
 use crate::{
 	circuits::slice::Slice,
-	compiler::{CircuitBuilder, Wire, WitnessFiller},
+	compiler::{CircuitBuilder, Wire, circuit::WitnessFiller},
 	word::Word,
 };
