+++ crates/frontend/src/circuits/sha256/mod.rs
@@ -3,7 +3,7 @@ pub mod compress;
 pub use compress::{Compress, State};
 
 use crate::{
-	compiler::{CircuitBuilder, Wire, WitnessFiller},
+	compiler::{CircuitBuilder, Wire, circuit::WitnessFiller},
 	word::Word,
 };
