+++ crates/frontend/src/compiler/gate/assert_eq.rs
@@ -13,11 +13,16 @@
 /// - `(x ⊕ y) ∧ all-1 = 0`
 use super::{Gate, GateData};
 use crate::{
-	compiler::{Circuit, WitnessFiller},
+	compiler::circuit,
 	constraint_system::{AndConstraint, ConstraintSystem},
 };
 
-pub fn constrain(_gate: Gate, data: &GateData, circuit: &Circuit, cs: &mut ConstraintSystem) {
+pub fn constrain(
+	_gate: Gate,
+	data: &GateData,
+	circuit: &circuit::Circuit,
+	cs: &mut ConstraintSystem,
+) {
 	let [x, y, all_1] = data.inputs() else {
 		unreachable!()
 	};
@@ -34,7 +39,7 @@ pub fn evaluate(
 	_gate: Gate,
 	data: &GateData,
 	assertion_name: Option<&String>,
-	w: &mut WitnessFiller,
+	w: &mut circuit::WitnessFiller,
 ) {
 	let [x, y, _all_1] = data.inputs() else {
 		unreachable!()
