+++ crates/frontend/src/compiler/gate/assert_eq_cond.rs
@@ -14,12 +14,17 @@
 /// - `(x ⊕ y) ∧ mask = 0`
 use super::{Gate, GateData};
 use crate::{
-	compiler::{Circuit, WitnessFiller},
+	compiler::circuit,
 	constraint_system::{AndConstraint, ConstraintSystem},
 	word::Word,
 };
 
-pub fn constrain(_gate: Gate, data: &GateData, circuit: &Circuit, cs: &mut ConstraintSystem) {
+pub fn constrain(
+	_gate: Gate,
+	data: &GateData,
+	circuit: &circuit::Circuit,
+	cs: &mut ConstraintSystem,
+) {
 	let [x, y, mask] = data.inputs() else {
 		unreachable!()
 	};
@@ -36,7 +41,7 @@ pub fn evaluate(
 	_gate: Gate,
 	data: &GateData,
 	assertion_name: Option<&String>,
-	w: &mut WitnessFiller,
+	w: &mut circuit::WitnessFiller,
 ) {
 	let [x, y, mask] = data.inputs() else {
 		unreachable!()
