+++ crates/frontend/src/compiler/gate/bor.rs
@@ -13,11 +13,16 @@
 /// - `x ∧ y = x ⊕ y ⊕ z`
 use super::{Gate, GateData};
 use crate::{
-	compiler::{Circuit, WitnessFiller},
+	compiler::circuit,
 	constraint_system::{AndConstraint, ConstraintSystem},
 };
 
-pub fn constrain(_gate: Gate, data: &GateData, circuit: &Circuit, cs: &mut ConstraintSystem) {
+pub fn constrain(
+	_gate: Gate,
+	data: &GateData,
+	circuit: &circuit::Circuit,
+	cs: &mut ConstraintSystem,
+) {
 	let [x, y] = data.inputs() else {
 		unreachable!()
 	};
@@ -33,7 +38,7 @@ pub fn constrain(_gate: Gate, data: &GateData, circuit: &Circuit, cs: &mut Const
 	cs.add_and_constraint(AndConstraint::plain_abc([x_idx], [y_idx], [x_idx, y_idx, z_idx]));
 }
 
-pub fn evaluate(_gate: Gate, data: &GateData, w: &mut WitnessFiller) {
+pub fn evaluate(_gate: Gate, data: &GateData, w: &mut circuit::WitnessFiller) {
 	let [x, y] = data.inputs() else {
 		unreachable!()
 	};
