+++ crates/frontend/src/compiler/gate/bxor.rs
@@ -13,11 +13,16 @@
 /// - `(x ⊕ y) ∧ all-1 = z`
 use super::{Gate, GateData};
 use crate::{
-	compiler::{Circuit, WitnessFiller},
+	compiler::circuit,
 	constraint_system::{AndConstraint, ConstraintSystem},
 };
 
-pub fn constrain(_gate: Gate, data: &GateData, circuit: &Circuit, cs: &mut ConstraintSystem) {
+pub fn constrain(
+	_gate: Gate,
+	data: &GateData,
+	circuit: &circuit::Circuit,
+	cs: &mut ConstraintSystem,
+) {
 	let [x, y, all_1] = data.inputs() else {
 		unreachable!()
 	};
@@ -34,7 +39,7 @@ pub fn constrain(_gate: Gate, data: &GateData, circuit: &Circuit, cs: &mut Const
 	cs.add_and_constraint(AndConstraint::plain_abc([x_idx, y_idx], [all_1_idx], [z_idx]));
 }
 
-pub fn evaluate(_gate: Gate, data: &GateData, w: &mut WitnessFiller) {
+pub fn evaluate(_gate: Gate, data: &GateData, w: &mut circuit::WitnessFiller) {
 	let [x, y, _all_1] = data.inputs() else {
 		unreachable!()
 	};
